SnapshotManager: Fix bug with optimizer tensors being on the wrong device
Optimizer has to be loaded after the model is already on the correct device, to
avoid a mismatch error.  So require the caller of load_snapshot to pass their
device.

Also fixed a bug where the wrong loss metric was being stored in the snapshot
metadata (loss of the last batch, not average loss from the epoch).