fetch_datasets: Use boto3 for s3 download instead of aws cli
Previous I was executing aws CLI as a subprocess so that I could get the nice
progress bar to stdout.  However, since this script is invoked by the ec2
userdata script, it's output goes to the cloud-init process and it was causing
kernel cpu lockup error and occasional outright crashes.

Going to go back to using the quieter method of download with no progress bar.