PerformanceTracker: Cache get_metrics().
On the large training dataset, there are 443k examples, and calculating the
performance metrics was taking a couple of seconds.  It turns out we call
get_metrics() at least 3 times after each snapshot epoch, so adding a cache will save 4
seconds (which is significant since the epoch training time is only 4 secs).