web: Support displaying of runs from 'mini' dataset
Previously we were excluding runs that were trained with --use-mini-dataset as
they are used for local testing on a tiny dataset.  However, it can be useful to
have mini runs on the website for locally testing end-to-end from a run to the
web interface.  So the website will show mini runs when ?useMini=true is passed
as a URL parameter.