Refactor embeddings handling out from VQADataset to EmbeddingsManager.
Also introducing configuration settings for the bert and vit model names.

EmbeddingsManager stores the embeddings in .pt files that are unique to the
model names, so if we change to a new embedding source, the embeddings will be
re-generated next run from the new source.

This will allow experimentation with different models for embeddings, such as
vit-huge, bert-large, and potentially others.