model: Support different text and image embedding sizes.
This will allow us to experiment with alternative text and image models that
might not necessarily have the same embedding size.

Also note that in this change, we weight each embedding type separately.  In the
previous version, the two weightings were normalised so that the weightings
summed to 1 at each feature index.  That created a bit of an arbitrary
competition between features in the image vs text, which could unnecessarily
limit their contributions.  With the new version, the weights don't have to sum
to 1, which should hopefully be beneficial.